requirements.txt
beautifulsoup4==4.12.2
Flask==2.2.3
flask_sqlalchemy==3.0.3
numpy==1.24.2
python-dotenv==1.0.0
requests==2.25.1

